# docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Restaurant Management System API
  version: 0.1.0
  description: Proposed MVP contract for the Express/TypeScript implementation. Slice 001 validates the document; each operation becomes implementation-authoritative when its delivery slice adds matching server behavior and contract tests.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /api
...
paths:
...
  /v1/auth/login:
    post:
      tags: [Identity]
      operationId: loginStaff
      summary: Create a server-managed, revocable staff session.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [businessCode, email, password]
              properties:
                businessCode: {type: string, minLength: 3, maxLength: 64, pattern: "^[a-z0-9][a-z0-9-]+$"}
                email: {type: string, format: email, maxLength: 320}
                password: {type: string, minLength: 1, maxLength: 1000, writeOnly: true}
      responses:
        "201":
          description: Authenticated; secure session and CSRF cookies are set.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/StaffSession"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "422": {$ref: "#/components/responses/ValidationFailed"}
        "429": {$ref: "#/components/responses/RateLimited"}
  /v1/auth/session:
    get:
      tags: [Identity]
      operationId: getStaffSession
      summary: Resolve the current revocable staff session.
      security: [{staffCookie: []}]
      responses:
        "200":
          description: Current session context.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/StaffSession"}
        "401": {$ref: "#/components/responses/Unauthorized"}
  /v1/auth/logout:
    post:
      tags: [Identity]
      operationId: logoutStaff
      summary: Revoke the current staff session.
      security: [{staffCookie: [], csrfHeader: []}]
      responses:
        "204": {description: Session revoked and cookies cleared.}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
  /v1/auth/recovery-requests:
    post:
      tags: [Identity]
      operationId: requestStaffCredentialRecovery
      summary: Request credential recovery without disclosing account existence.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [businessCode, email]
              properties:
                businessCode: {type: string, minLength: 3, maxLength: 64}
                email: {type: string, format: email, maxLength: 320}
      responses:
        "202": {description: Request accepted regardless of account existence.}
        "422": {$ref: "#/components/responses/ValidationFailed"}
        "429": {$ref: "#/components/responses/RateLimited"}
        "503": {$ref: "#/components/responses/ServiceUnavailable"}
  /v1/auth/recovery-completions:
    post:
      tags: [Identity]
      operationId: completeStaffCredentialRecovery
      summary: Consume a recovery token, replace the credential, and revoke existing sessions.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/TokenAndPassword"}
      responses:
        "204": {description: Credential replaced and sessions revoked.}
        "404": {$ref: "#/components/responses/NotFound"}
        "422": {$ref: "#/components/responses/ValidationFailed"}
        "429": {$ref: "#/components/responses/RateLimited"}
  /v1/invitations/accept:
    post:
      tags: [Identity]
      operationId: acceptStaffInvitation
      summary: Consume a single-use invitation without adding grants.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/TokenAndPassword"}
      responses:
        "204": {description: Invitation accepted.}
        "404": {$ref: "#/components/responses/NotFound"}
        "422": {$ref: "#/components/responses/ValidationFailed"}
  /v1/staff/session/branch:
    post:
      tags: [Identity]
      operationId: switchStaffBranch
      summary: Switch to a branch assigned to the current employee.
      security: [{staffCookie: [], csrfHeader: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [branchId]
              properties:
                branchId: {type: string, format: uuid}
      responses:
        "200":
          description: Active branch changed.
          content:
            application/json:
              schema:
                type: object
                required: [activeBranchId]
                properties:
                  activeBranchId: {type: string, format: uuid}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
...
  /v1/staff/employees/{employeeId}/invitations:
    post:
      tags: [Identity]
      operationId: inviteExistingEmployee
      summary: Create a single-use invitation for an existing employee profile.
      security: [{staffCookie: [], csrfHeader: []}]
      parameters:
        - $ref: "#/components/parameters/EmployeeId"
      responses:
        "201":
          description: Raw invitation token returned once to the authorized administrator.
          content:
            application/json:
              schema:
                type: object
                required: [invitationToken, expiresAt]
                properties:
                  invitationToken: {type: string, writeOnly: true}
                  expiresAt: {type: string, format: date-time}
        "404": {$ref: "#/components/responses/NotFound"}
  /v1/staff/administrators/{employeeId}/removal:
    post:
      tags: [Identity]
      operationId: removeAdministrator
      summary: Remove administrator grants unless this would remove the final effective administrator.
      security: [{staffCookie: [], csrfHeader: []}]
      parameters:
        - $ref: "#/components/parameters/EmployeeId"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/Reason"}
      responses:
        "204": {description: Administrator grants removed and sessions revoked.}
        "409": {$ref: "#/components/responses/Conflict"}
  /v1/staff/administrators/transfer:
    post:
      tags: [Identity]
      operationId: transferAdministrator
      summary: Activate a replacement administrator before optionally removing the current administrator.
      security: [{staffCookie: [], csrfHeader: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [replacementEmployeeId, removeCurrentAdministrator, reason]
              properties:
                replacementEmployeeId: {type: string, format: uuid}
                removeCurrentAdministrator: {type: boolean}
                reason: {type: string, minLength: 8, maxLength: 500}
      responses:
        "204": {description: Replacement activated and transfer audited.}
        "409": {$ref: "#/components/responses/Conflict"}
  /v1/staff/employees/{employeeId}/deactivation:
    post:
      tags: [Identity]
      operationId: deactivateEmployeeIdentity
      summary: Atomically deactivate an employee identity and revoke sessions, subject to the final-administrator guard.
      security: [{staffCookie: [], csrfHeader: []}]
      parameters:
        - $ref: "#/components/parameters/EmployeeId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/Reason"
                - type: object
                  required: [expectedVersion]
                  properties:
                    expectedVersion: {type: integer, minimum: 1}
      responses:
        "204": {description: Employee and identity deactivated; sessions revoked.}
        "409": {$ref: "#/components/responses/Conflict"}
...
components:
  securitySchemes:
    bootstrapSecret:
      type: apiKey
      in: header
      name: X-Bootstrap-Secret
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
    staffCookie:
      type: apiKey
      in: cookie
      name: rms_staff_session
    guestCookie:
      type: apiKey
      in: cookie
      name: rms_guest_session
  parameters:
...
    EmployeeId:
      name: employeeId
      in: path
      required: true
      schema: {type: string, format: uuid}
...
  responses:
    Unauthorized:
      description: Authentication required.
      content: {application/problem+json: {schema: {$ref: "#/components/schemas/Problem"}}}
    Forbidden:
      description: Authenticated but not authorized.
      content: {application/problem+json: {schema: {$ref: "#/components/schemas/Problem"}}}
    NotFound:
      description: Resource absent or not visible in the current scope.
      content: {application/problem+json: {schema: {$ref: "#/components/schemas/Problem"}}}
    Conflict:
      description: State, version, menu, table, payment, or idempotency conflict.
      content: {application/problem+json: {schema: {$ref: "#/components/schemas/Problem"}}}
    ValidationFailed:
      description: Request validation failed.
      content: {application/problem+json: {schema: {$ref: "#/components/schemas/Problem"}}}
    RateLimited:
      description: Rate limit exceeded.
      headers:
        Retry-After: {schema: {type: integer}}
      content: {application/problem+json: {schema: {$ref: "#/components/schemas/Problem"}}}
    ServiceUnavailable:
      description: A required delivery dependency is unavailable.
      content: {application/problem+json: {schema: {$ref: "#/components/schemas/Problem"}}}
  schemas:
    StaffSession:
      type: object
      required: [employeeId, activeBranchId, expiresAt]
      properties:
        employeeId: {type: string, format: uuid}
        activeBranchId: {type: [string, "null"], format: uuid}
        authorizedBranchIds: {type: array, items: {type: string, format: uuid}}
        expiresAt: {type: string, format: date-time}
    TokenAndPassword:
      type: object
      required: [token, password]
      properties:
        token: {type: string, minLength: 32, maxLength: 256, writeOnly: true}
        password: {type: string, minLength: 12, maxLength: 128, writeOnly: true}
    Reason:
      type: object
      required: [reason]
      properties:
        reason: {type: string, minLength: 8, maxLength: 500}
...
    Problem:
      type: object
      required: [type, title, status, code, correlationId]
      properties:
        type: {type: string, format: uri-reference}
        title: {type: string}
        status: {type: integer}
        code:
          type: string
          enum: [validation_error, authentication_required, permission_denied, feature_disabled, resource_not_found, invalid_state_transition, concurrency_conflict, menu_changed, dish_unavailable, table_unavailable, payment_conflict, idempotency_conflict, rate_limited, service_unavailable, unexpected_error]
        detail: {type: [string, "null"]}
        correlationId: {type: string}
        currentVersion: {type: integer, minimum: 1}
        errors:
          type: object
          additionalProperties:
            type: array
            items: {type: string}